+++ spec/helpers/src/MockResourceFactory.ts
@@ -1,13 +1,14 @@
 import * as fs from "fs";
 import * as path from "path";
-import * as g from "@akashic/akashic-engine";
+import * as pdi from "@akashic/pdi-types";
+import * as pci from "@akashic/pdi-common-impl";
 
 export interface MethodCallParam {
 	methodName: string;
 	params?: any;
 }
 
-export class Renderer extends g.Renderer {
+export class Renderer extends pci.Renderer {
 	methodCallHistoryWithParams: MethodCallParam[];
 
 	constructor() {
@@ -41,7 +42,7 @@ export class Renderer extends g.Renderer {
 		return params;
 	}
 
-	drawImage(surface: g.Surface, offsetX: number, offsetY: number, width: number, height: number,
+	drawImage(surface: pdi.Surface, offsetX: number, offsetY: number, width: number, height: number,
 	          canvasOffsetX: number, canvasOffsetY: number): void {
 		this.methodCallHistoryWithParams.push({
 			methodName: "drawImage",
@@ -85,7 +86,7 @@ export class Renderer extends g.Renderer {
 		});
 	}
 
-	setCompositeOperation(operation: g.CompositeOperationString): void {
+	setCompositeOperation(operation: pdi.CompositeOperationString): void {
 		this.methodCallHistoryWithParams.push({
 			methodName: "setCompositeOperation",
 			params: {
@@ -119,7 +120,7 @@ export class Renderer extends g.Renderer {
 		});
 	}
 
-	drawSprites(surface: g.Surface,
+	drawSprites(surface: pdi.Surface,
 	            offsetX: number[], offsetY: number[],
 	            width: number[], height: number[],
 	            canvasOffsetX: number[], canvasOffsetY: number[],
@@ -139,28 +140,6 @@ export class Renderer extends g.Renderer {
 		});
 	}
 
-	drawSystemText(text: string, x: number, y: number, maxWidth: number, fontSize: number,
-	               textAlign: g.TextAlign, textBaseline: g.TextBaseline, textColor: string, fontFamily: g.FontFamily,
-	               strokeWidth: number, strokeColor: string, strokeOnly: boolean): void {
-		this.methodCallHistoryWithParams.push({
-			methodName: "drawSystemText",
-			params: {
-				text: text,
-				x: x,
-				y: y,
-				maxWidth: maxWidth,
-				fontSize: fontSize,
-				textAlign: textAlign,
-				textBaseline: textBaseline,
-				textColor: textColor,
-				fontFamily: fontFamily,
-				strokeWidth: strokeWidth,
-				strokeColor: strokeColor,
-				strokeOnly: strokeOnly
-			}
-		});
-	}
-
 	isSupportedShaderProgram(): boolean {
 		throw new Error("not implemented: mock renderer isSupportedShaderProgram()");
 	}
@@ -169,7 +148,7 @@ export class Renderer extends g.Renderer {
 		throw new Error("not implemented: mock renderer setOpacity()");
 	}
 
-	setShaderProgram(shaderProgram: g.ShaderProgram | null): void {
+	setShaderProgram(shaderProgram: pdi.ShaderProgram | null): void {
 		throw new Error("not implemented: mock renderer setShaderProgram()");
 	}
 
@@ -177,23 +156,23 @@ export class Renderer extends g.Renderer {
 		throw new Error("not implemented: mock renderer setTransform()");
 	}
 
-	_getImageData(sx: number, sy: number, sw: number, sh: number): g.ImageData {
+	_getImageData(sx: number, sy: number, sw: number, sh: number): pdi.ImageData {
 		throw new Error("not implemented: mock renderer _getImageData()");
 	}
 
-	_putImageData(imageData: g.ImageData, dx: number, dy: number, dw: number, dh: number): void {
+	_putImageData(imageData: pdi.ImageData, dx: number, dy: number, dw: number, dh: number): void {
 		throw new Error("not implemented: mock renderer _putImageData()");
 	}
 }
 
-class Surface extends g.Surface {
-	createdRenderer: g.Renderer;
+class Surface extends pci.Surface {
+	createdRenderer: pdi.Renderer;
 
 	constructor(width: number, height: number, drawable?: any) {
 		super(width, height, drawable);
 	}
 
-	renderer(): g.Renderer {
+	renderer(): pdi.Renderer {
 		var r = new Renderer();
 		this.createdRenderer = r;
 		return r;
@@ -213,18 +192,18 @@ export class LoadFailureController {
 		this.failureCount = 0;
 	}
 
-	tryLoad(asset: g.Asset, loader: g.AssetLoadHandler): boolean {
+	tryLoad(asset: pdi.Asset, loader: pdi.AssetLoadHandler): boolean {
 		if (this.necessaryRetryCount < 0) {
 			setTimeout(() => {
 				if (!asset.destroyed())
-					loader._onAssetError(asset, g.ExceptionFactory.createAssetLoadError("FatalErrorForAssetLoad", false));
+					loader._onAssetError(asset, { name: "AssetLoadError", message: "FatalErrorForAssetLoad", retriable: false });
 			}, 0);
 			return false;
 		}
 		if (this.failureCount++ < this.necessaryRetryCount) {
 			setTimeout(() => {
 				if (!asset.destroyed())
-					loader._onAssetError(asset, g.ExceptionFactory.createAssetLoadError("RetriableErrorForAssetLoad"));
+					loader._onAssetError(asset, { name: "AssetLoadError", message: "RetriableErrorForAssetLoad", retriable: true });
 			}, 0);
 			return false;
 		}
@@ -232,7 +211,7 @@ export class LoadFailureController {
 	}
 }
 
-export class ImageAsset extends g.ImageAsset {
+export class ImageAsset extends pci.ImageAsset {
 	_failureController: LoadFailureController;
 
 	constructor(necessaryRetryCount: number, id: string, assetPath: string, width: number, height: number) {
@@ -240,7 +219,7 @@ export class ImageAsset extends g.ImageAsset {
 		this._failureController = new LoadFailureController(necessaryRetryCount);
 	}
 
-	_load(loader: g.AssetLoadHandler): void {
+	_load(loader: pdi.AssetLoadHandler): void {
 		if (this._failureController.tryLoad(this, loader)) {
 			setTimeout(() => {
 				if (!this.destroyed())
@@ -249,21 +228,21 @@ export class ImageAsset extends g.ImageAsset {
 		}
 	}
 
-	asSurface(): g.Surface {
+	asSurface(): pdi.Surface {
 		return new Surface(0, 0);
 	}
 }
 
-class AudioAsset extends g.AudioAsset {
+class AudioAsset extends pci.AudioAsset {
 	_failureController: LoadFailureController;
 
 	constructor(necessaryRetryCount: number, id: string, assetPath: string, duration: number,
-	            system: g.AudioSystem, loop: boolean, hint: g.AudioAssetHint) {
+	            system: pdi.AudioSystem, loop: boolean, hint: pdi.AudioAssetHint) {
 		super(id, assetPath, duration, system, loop, hint);
 		this._failureController = new LoadFailureController(necessaryRetryCount);
 	}
 
-	_load(loader: g.AssetLoadHandler): void {
+	_load(loader: pdi.AssetLoadHandler): void {
 		if (this._failureController.tryLoad(this, loader)) {
 			setTimeout(() => {
 				if (!this.destroyed())
@@ -273,7 +252,7 @@ class AudioAsset extends g.AudioAsset {
 	}
 }
 
-class TextAsset extends g.TextAsset {
+class TextAsset extends pci.TextAsset {
 	resourceFactory: ResourceFactory;
 	_failureController: LoadFailureController;
 
@@ -283,7 +262,7 @@ class TextAsset extends g.TextAsset {
 		this._failureController = new LoadFailureController(necessaryRetryCount);
 	}
 
-	_load(loader: g.AssetLoadHandler): void {
+	_load(loader: pdi.AssetLoadHandler): void {
 		if (this._failureController.tryLoad(this, loader)) {
 			setTimeout(() => {
 				if ((this.resourceFactory).scriptContents.hasOwnProperty(this.path)) {
@@ -298,7 +277,7 @@ class TextAsset extends g.TextAsset {
 	}
 }
 
-class ScriptAsset extends g.ScriptAsset {
+class ScriptAsset extends pci.ScriptAsset {
 	resourceFactory: ResourceFactory;
 	_failureController: LoadFailureController;
 	_content: string;
@@ -310,11 +289,11 @@ class ScriptAsset extends g.ScriptAsset {
 		this._content = null;
 	}
 
-	_load(loader: g.AssetLoadHandler): void {
+	_load(loader: pdi.AssetLoadHandler): void {
 		if (this._failureController.tryLoad(this, loader)) {
 			fs.readFile( path.resolve(this.path), "utf8", (err: any, data: string) => {
 				if (err) {
-					loader._onAssetError(this, g.ExceptionFactory.createAssetLoadError("FatalErrorForAssetLoad", false));
+					loader._onAssetError(this, { name: "AssetLoadError", message: "FatalErrorForAssetLoad", retriable: false });
 					return;
 				}
 				this._content = data;
@@ -324,7 +303,7 @@ class ScriptAsset extends g.ScriptAsset {
 		}
 	}
 
-	execute(env: g.ScriptAssetRuntimeValue): any {
+	execute(env: pdi.ScriptAssetRuntimeValue): any {
 		var prefix = "(function(exports, require, module, __filename, __dirname) {";
 		var suffix = "})(g.module.exports, g.module.require, g.module, g.filename, g.dirname);";
 		var f = new Function("g", prefix + this._content + suffix);
@@ -333,11 +312,11 @@ class ScriptAsset extends g.ScriptAsset {
 	}
 }
 
-export class AudioPlayer extends g.AudioPlayer {
+export class AudioPlayer extends pci.AudioPlayer {
 	supportsPlaybackRateValue: boolean;
 	canHandleStoppedValue: boolean;
 
-	constructor(system: g.AudioSystem) {
+	constructor(system: pdi.AudioSystem) {
 		super(system);
 		this.supportsPlaybackRateValue = true;
 		this.canHandleStoppedValue = true;
@@ -352,7 +331,7 @@ export class AudioPlayer extends g.AudioPlayer {
 	}
 }
 
-export class ResourceFactory extends g.ResourceFactory {
+export class ResourceFactory extends pci.ResourceFactory {
 	scriptContents: {[key: string]: string};
 
 	_necessaryRetryCount: number;
@@ -376,39 +355,39 @@ export class ResourceFactory extends g.ResourceFactory {
 		}
 	}
 
-	createImageAsset(id: string, assetPath: string, width: number, height: number): g.ImageAsset {
+	createImageAsset(id: string, assetPath: string, width: number, height: number): ImageAsset {
 		return new ImageAsset(this._necessaryRetryCount, id, assetPath, width, height);
 	}
 
 	createAudioAsset(id: string, assetPath: string, duration: number,
-	                 system: g.AudioSystem, loop: boolean, hint: g.AudioAssetHint): g.AudioAsset {
+	                 system: pdi.AudioSystem, loop: boolean, hint: pdi.AudioAssetHint): AudioAsset {
 		return new AudioAsset(this._necessaryRetryCount, id, assetPath, duration, system, loop, hint);
 	}
 
-	createTextAsset(id: string, assetPath: string): g.TextAsset {
+	createTextAsset(id: string, assetPath: string): TextAsset {
 		return new TextAsset(this, this._necessaryRetryCount, id, assetPath);
 	}
 
-	createScriptAsset(id: string, assetPath: string): g.ScriptAsset {
+	createScriptAsset(id: string, assetPath: string): ScriptAsset {
 		return new ScriptAsset(this, this._necessaryRetryCount, id, assetPath);
 	}
 
-	createSurface(width: number, height: number): g.Surface {
+	createSurface(width: number, height: number): Surface {
 		return new Surface(width, height);
 	}
 
-	createAudioPlayer(system: g.AudioSystem): g.AudioPlayer {
+	createAudioPlayer(system: pdi.AudioSystem): AudioPlayer {
 		return new AudioPlayer(system);
 	}
 
 	createVideoAsset(id: string, assetPath: string, width: number, height: number,
-	                 system: g.VideoSystem, loop: boolean, useRealSize: boolean): g.VideoAsset {
+	                 system: pdi.VideoSystem, loop: boolean, useRealSize: boolean): pci.VideoAsset {
 		throw new Error("not implemented: mock resourceFactory createVideoAsset()");
 	}
 
 	createGlyphFactory(fontFamily: string | string[], fontSize: number,
 	                   baselineHeight?: number, fontColor?: string, strokeWidth?: number,
-	                   strokeColor?: string, strokeOnly?: boolean, fontWeight?: g.FontWeightString): g.GlyphFactory {
+	                   strokeColor?: string, strokeOnly?: boolean, fontWeight?: pdi.FontWeightString): pci.GlyphFactory {
 		throw new Error("not implemented: mock resourceFactory createGlyphFactory()");
 	}
 }