+++ spec/src/EventBufferSpec.ts
@@ -1,5 +1,6 @@
 import * as pl from "@akashic/playlog";
-import { EventIndex, EventPriority, PlatformPointEvent, PlatformPointType } from "@akashic/akashic-engine";
+import { PlatformPointEvent, PlatformPointType } from "@akashic/pdi-types";
+import { EventIndex, EventPriority } from "@akashic/akashic-engine";
 import { MockAmflow } from "../helpers/lib/MockAmflow";
 import { prepareGame, FixtureGame } from "../helpers/lib/prepareGame";
 import { EventBuffer } from "../../lib/EventBuffer";